pyproject.toml
[project]
name = "pavlov-bot"
version = "2.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.16",
    "async-pavlov>=0.4.0",
    "beautifulsoup4>=4.13.3",
    "discord-py>=2.5.2",
    "pillow>=11.1.0",
]
